Configuração de Adapters + Models + Rename of Application

+++ app/src/main/java/br/projeto/apanhagastos/models/Gasto.kt
@@ -3,5 +3,5 @@ package br.projeto.apanhagastos.models
 data class Gasto (
     val nomeGasto: String = "",
     val categoria: String = "",
-    val custo: Long = 0L,
+    val custo: Double = 0.00,
 )
