+++ app/src/main/java/br/projeto/apanhagastos/models/GastoComId.kt
@@ -3,6 +3,6 @@ package br.projeto.apanhagastos.models
 data class GastoComId (
     val nomeGasto: String = "",
     val categoria: String = "",
-    val custo: Long = 0L,
+    val custo: Double = 0.00,
     val id: String = ""
 )
