+++ app/src/main/java/br/projeto/apanhagastos/models/GastoEmUsuario.kt
@@ -3,7 +3,7 @@ package br.projeto.apanhagastos.models
 data class GastoEmUsuario(
     val nomeGasto: String = "",
     val categoria: String = "",
-    val custo: Long = 0L,
+    val custo: Double = 0.00,
 
     val nomeUsuario: String = ""
 )
